+++ day04/lib/aht20.h
@@ -2,12 +2,12 @@
 
 typedef struct
 {
-	uint64_t humidity;
-	uint64_t temp;
+	uint32_t humidity;
+	uint32_t temp;
 } aht20_data;
 
 void aht20_init(void);
 aht20_data aht20_mesure(void);
 
-void print_temp(uint64_t temp);
-void print_humi(uint64_t humi);
+void print_temp(uint32_t temp);
+void print_humi(uint32_t humi);
